Changed to be compatible with 7.0 server source
https://github.com/dhojka7/realm-server

+++ Assets/Scripts/Networking/Data/ObjectStatusData.cs
@@ -13,7 +13,7 @@
             pos = new WorldPosData();
             pos.Read(nR);
 
-            stats = new StatData[nR.ReadInt16()];
+            stats = new StatData[nR.ReadByte()];
             for (int i = 0; i < stats.Length; i++)
             {
                 stats[i] = new StatData();
